requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.9.8

# Configuration
django-environ==0.4.0
whitenoise==3.2


# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.3
redis>=2.10.0







django-rest-swagger==2.0.2
djangorestframework==3.4.0

# sqlparse 0.2 is incompatible with django_debug_toolbar
sqlparse==0.1.19 # pyup: < 0.2

# Your custom requirements go here
